requirements.txt
PyQt6>=6.5.0
requests>=2.31.0
markdown>=3.4.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
# QMarkdownTextEdit: если требуется, установите вручную из https://github.com/pbek/QMarkdownTextEdit
# PyQt6 QTextEdit имеет встроенную поддержку Markdown через setMarkdown/toMarkdown
